Add url\_launcher package to enable launching URLs in Flutter app
This commit adds the url\_launcher package to the project to enable launching URLs from within the Flutter app. The package is added to the pubspec.yaml file and its dependencies are also added. The necessary changes are made in the AndroidManifest.xml and Info.plist files to allow querying of certain URL schemes. The code is updated to use the UrlLauncher package to launch URLs for free preview of books. The plugin is also registered for each platform in their respective registrant files.